[runtime] Create initial bytecode VM
Initial commit for the new bytecode VM, which will replace the tree walk interpreter. This bytecode VM implements a subset of the language in order to compile and execute a test fibonacci function.

Highlights include:

- Variable width bytecode, with wide, narrow, and extra wide instructions
- Register based bytecode VM
- Single pass bytecode generation from AST
  - Jump patching using reserved constants
  - Constant table generation
  - Simple temporary register allocation (stack based)
- OpCodes generated from custom DSL
- Bytecode generated in managed heap objects, compilation and execution is GC aware

The only known issue in execution of the current bytecode VM is that it does not rewrite the VM stack during a GC - coming soon.