[bytecode] Remove handle pool from VM to fix re-entrant function calls
The VM handle pool was not safe as an instruction that used the handle pool could make a re-entrant call to the VM from the rust runtime, clobbering the handle pool (e.g. in the add instruction). Instead we must allocate a new handle each time we need one in the VM.