[parser][bytecode] Avoid assignment hazards when generating bytecode
Avoid miscompiles due to assignment hazards e.g. `x + (x = 1)`. Take a conservative approach, where if we are ever in an expression that contains an assignment expression, all references to fixed registers (i.e. locals and parameters) must be loaded to a temporary register instead of directly referencing the fixed register.

In the AST we ensure that every expression is part of a single "has assignment expression" context. Expressions held by statements are wrapped in the OuterExpression node. Any statement that holds a pattern also must hold a "has assignment expression" flag for that pattern due to computed pattern properties or assignment patterns.

During analysis we fill the "has assignment expression" flag for each context. Then during generation we check this flag when evaluating an identifier to see if we need to load to a temporary.