[analyze] Arguments object necessity is based on scope analysis
Whether an arguments object is needed is now based off scope analysis instead of naively needing the arguments object whenever the identifier "arguments" appears.

An implicit arguments object binding is added to function scopes during analysis when trying to resolve a use of "arguments" that has not found a binding.

We treat both implicit arguments, var, and var function bindings with the name "arguments" as potential arguments objects. This is because it is hard to accurately rule out the necessity of the arguments object for these var scoped bindings, as they would have the value of the arguments object before they are initialized.

Any "force VM scope" will also force a function to create an implicit arguments object.