[bytecode] Generate bytecode for eval return value, storing result of expression statements, create eval test harness
We now generate bytecoe for eval's return value. The return value is the value of the most recently executed expression statement, which we store in a known register and return at the end of eval.

This commit also sets up a test harness for eval bytecode snapshot tests. This is enabled with the new `// OPTIONS: --run` syntax at the top of bytecode snapshot tests that should be run. And the eval bytecode itself is written to the internal dump buffer which has been added to the shared options.