[bytecode] Generate bytecode for finally statements, including returns in a finally scope
Finally statements are generated once, but can have multiple predecessors and successors due to the variety of control flow branches (e.g. normal, throw, return, break, continue) that must flow through them.

Enter a finally scope when generating the body of a finally statement. When in this scope, all returns need to first jump to the body of the finally, then proceed with the return.

Finally blocks have a discriminant and shared result value. The descriminant is set by the incoming branch to mark whether it is a normal, throw, return, break, or continue. The shared result value is optionally set to the return value or exception to throw. At the end of the finally block we check the discriminant to see if we should continue normally, throw, return, or jump to a break/continue.