[bytecode] Sloppy direct eval errors when introducing var scoped names that conflict with names in parent scopes
This includes a special case for evals that occur in function parameter expressions. In this case a separate function body scope must always be created to distinguish params and body scopes. We then track whether names are function parameters, and only check these param names when performing a sloppy direct eval within the functin params scope. Evals in the function body will not ascend past the function body scope when checking for name conflicts.