[bytecode] Refactor global declaration instantiation, move to runtime, and check for conflicting lexical names
Global declaration instantiation is now a function in the rust runtime that is called before executing the script function. The GlobalNames object now contains information about global lexical names as well, and is created during bytecode gen and returned alongside the script function as a BytecodeProgram struct. The GlobalNames are then passed to GDI when called from the VM. This is always called, and replaces the GlobalInit instruction.

Global declaration instantiation also now checks for lexical names that conflict with existing global names.

Also includes related bugfixes:
- $262.evalScript now uses the bytecode VM
- GlobalNames correctly tracks which names are function vs vars