[bytecode] Generate bytecode for class fields
Class fields are generated as DefineProperty instructions either in the constructor (for instance fields) or on the constructor after creating the class (for static fields).

Class field keys must be computed in order along with the method keys. This means that computed field keys will be generated before the NewClass instruction and used later, potentially in the constructor, so they must be saved. Computed field keys are stored in the class body scope so that they can be accessed by the constructor and do not interfere with the contiguous sequence of registers for the NewClass instruction.