[bytecode] VM uses volatile reads and writes to PC
A recent commit exposed an issue with reads and writes to fields in the VM struct. Rust assumes single mutable ownership and was optimizing out reads of fields, even though those fields could have been rewritten by the GC. Enforcing that reads and writes to the affected fields are volatile fixes this.

The only affected field was the PC. However all VM field accesses are now wrapped in getters/setters for consistency and ease of refactoring in the future.