[ci] Check for clippy and rustfmt issues in CI (#1)
Ensure that we do not merge changes that introduce clippy or formatting issues by running clippy and rustfmt in CI. These jobs will run both on PRs and on merging to master.

Note this also implies a process change for this repo. All further changes should first go through a PR and only be merged to master if all CI jobs are passing.