[runtime] MapObject and SetObject iterators handle resizing the underlying collection (#11)
## Summary

MapObject and SetObject iterators maintain a direct reference to the
underlying BsIndexMap. However the map or set can be resized while
iterating, which creates a different BsIndexMap with a copy of the
collection's contents. But the iterator still has a reference to the old
underlying BsIndexMap.

To handle resizing we must rewrite the old BsIndexMap in place to
contain the information needed to update the iterator to point to the
new BsIndexMap. This old BsIndexMap is called a tombstone which we set a
flag to indicate.

The tombstone needs to contain the pointer to the new BsIndexMap, which
is stored at the first entry of the indices array. We also need
information to update the iterator's old next entry index. We can update
this by counting the number of occupied entries in the old map up to
that index. Note that we only need the occupied flag, so tombstones do
not hold strong references to their entries.

Also fixes a GC safety bug when inserting into BsIndexMap and BsIndexSet
with ValueCollectionKeys.