[runtime] Fix GC safety bugs in typed arrays
Fixes two instances of the same class of GC safety bug in typed arrays. When copying values from a typed array to another of a different type, we were holding a raw pointer over an allocation. This was occurring in the typed array constructor as well as TypedArray.prototype.set.

- Now passes test262 tests with GC stress test mode enabled (`cargo run --release --features gc_stress_test`)