[runtime] Fix typed array [[DefineOwnProperty]] and [[Set]] bug caused be shrinking ArrayBuffer
Fixes a bug in typed array [[DefineOwnProperty]] and [[Set]]. Converting the input value into the typed array's expected type can invoke user code, which may resize the underlying ArrayBuffer. If the index is initially out of bounds but user code is invoked to shrink it so that the index is in range, we should use the new index. Previously we crashed on this case.