[runtime] Fix typed array [[DefineOwnProperty]] and [[Set]] bug caused be shrinking ArrayBuffer (#17)
## Summary

Fixes a bug in typed array [[DefineOwnProperty]] and [[Set]]. Converting
the input value into the typed array's expected type can invoke user
code, which may resize the underlying ArrayBuffer. If the index is
initially out of bounds but user code is invoked to shrink it so that
the index is in range, we should use the new index. Previously we
crashed on this case.

## Tests

-
[built-ins/TypedArray/of/resized-with-out-of-bounds-and-in-bounds-indices.js](https://github.com/tc39/test262/blob/98c7e9114e7eb663f895953c3750c6b270660a51/test/built-ins/TypedArray/of/resized-with-out-of-bounds-and-in-bounds-indices.js)
-
[built-ins/TypedArrayConstructors/internals/Set/resized-out-of-bounds-to-in-bounds-index.js](https://github.com/tc39/test262/blob/98c7e9114e7eb663f895953c3750c6b270660a51/test/built-ins/TypedArrayConstructors/internals/Set/resized-out-of-bounds-to-in-bounds-index.js)