[runtime] Fix ArrayIterator.prototype.next when called on a resized TypedArray (#20)
## Summary

Fixes a bug in `ArrayIterator.prototype.next` when called on a resized
TypedArray. This exposed that iterators that have already completed
could throw an error on `next` during the call to `get_length`, e.g. if
the TypedArray has been resized making the current index out of bounds.
The solution is to immediately return from
`ArrayIterator.prototype.next` if the iterator is done before
potentially invoking any user code.

## Tests

- [built-ins/TypedArray/prototype/values/make-out-of-bounds-after-exhausted.js](https://github.com/tc39/test262/blob/98c7e9114e7eb663f895953c3750c6b270660a51/test/built-ins/TypedArray/prototype/values/make-out-of-bounds-after-exhausted.js)