[regexp] Always emit progress instructions in optional quantifier iterations that are not guaranteed to consume (#22)
## Summary

Fixes bugs concerning the progress instruction in quantifiers. The
progress instruction is used in optional iterations of quantifiers that
are not guaranteed to consume any characters. This ensures that forward
progress is made (avoiding infinite epsilon loops), and also makes it so
that optional quantifier iterations do not match if they match the empty
string.

Previously we were only emitting the progress instruction for
quantifiers with an uncapped max. We should also be emitting progress
instructions for quantifiers with fixed maximums.

Additionally we were checking progress on the back edge of the loop (and
were branching beforehand), which means the quantifier iteration could
match the empty string. Instead we should always be checking progress
right after the loop, but before the branch to either the beginning of
loop or end.

## Tests

- Fixes
[built-ins/RegExp/nullable-quantifier.js](https://github.com/tc39/test262/blob/867ca540d6cc991a82b41a3b7f9ccb9e93efe803/test/built-ins/RegExp/nullable-quantifier.js)