[runtime] Implement basic module loading (#29)
## Summary

Implements basic module loading and the beginnings of SourceTextModule.
Parsing modules now is followed by generating bytecode for that module,
which will return a SourceTextModule. This SourceTextModule can have all
its dependencies loaded following the spec's loading algorithm. The
later linking and evaluation phases are not yet implemented.

Also implements basic module execution. The entrypoint is
`Context::run_module` which will run the module, loading, linking, and
evaluating it and all its module dependencies. Note that these phases
are all async so encapsulate the entire execution in a promise, then
must specify the behavior when the promise rejects (print error and exit
vs panic).

Modules are loaded treating the module specifier as a path which may be
absolute or relative to the source file. The path to a source file is
canonicalized and the module at that path will only be loaded once and
the SourceTextModule is cached for later loads.