[analyze][bytecode] Analyze and generate bytecode for simple imported and exported bindings
Adds a new binding kind for imports and marks all exported bindings as exports. All imported and exported bindings have a new VM location where they are stored in a slot of the module scope.

Also adds the LoadFromModule and StoreInModule instructions for loading and storing a module binding. The value stored is a BoxedValue which is a simple heap wrapper for values.

Does not yet support `export *` declarations.