[runtime] Link imports with exports and initialize module scopes
Implements the link phase for modules. The module scope is now initialized with BoxedValues stored for every export (during loading). Imports are linked to exports during the link phase by referencing these same BoxedValues.

Note that namespace imports are normally stored as regular scope bindings, and are only stored as module bindings when exported.

Namespace imports are not yet implemented.