[runtime] Patch exported functions into the module scope (#34)
## Summary

Exported function bindings need to be initialized before the module
evaluation phase begins. This is implemented by directly patching the
closure for the exported function into the module scope once the
exported function has had its bytecode generated.

This is done by generalizing the types of patches that can be applied to
the pending functions after bytecode generation. We now support patching
a parent function's constant table and patching the module scope.