[test] Check whether errors occur during module resolution phase (#36)
## Summary

Fixes handling of module test errors in the test262 runner. We now check
that expected errors occur in the correct phase - module resolution or
runtime. This is accomplished with a flag on the Context which is set
after module resolution completes.

In addition we have also changed how errors during module resolution and
evaluation are handled in the top level `Context::run_module` function.
Previously we added a promise then reaction which either panics or
exists on promise rejection but this did not play well with the test262
runner. Instead we can check whether the toplevel promise for module
execution has rejected after all tasks have completed, and if so return
an error.