[runtime] SourceFile object has a lazily generated array of line offsets (#38)
## Summary

We will want line offsets stored in the heap so that they can be used to
determine source positions and extract lines of source text. Line
offsets are stored in a BsArray as a field of the SourceFile object. The
line offsets array is lazily generated when first requested.

Also generalizes the ArrayBuffer's DataArray to a ByteArray, which can
be used for any opaque data stored on the heap (such as an array of u32
line offsets).

Finally cleans up some unnecessary `#[allow(unused)]` attributes.