[analyze] Resolve export specifiers (#42)
## Summary

Export specifiers were not yet working because we were failing to
resolve the local identifier to its definition. With this fixed export
specifiers are now working.

Note that we should only resolve the local identifier when the export is
not a re-export.

Also adds bytecode snapshot tests for named export specifiers, including
re-exports.