[runtime] Fix dynamic lookup of module bindings (#46)
## Summary

Fixes dynamic lookup of module bindings which previously were
loading/storing the entire BoxedValue instead of the inner value. This
is accomplished by adding a scope name flag for module bindings and
checking this flag during dynamic lookup, treating value in slot as a
BoxedValue if the flag is set.

Note that we also introduce a ScopeKind::Module which also gates the
module binding check during dynamic lookup.