[parser] Allow meta properties and super properties as callee in new expression (#54)
## Summary

The callee in a new expression should be a MemberExpression, however we
were failing to parse meta properties and super properties even though
they are MemberExpressions. We should parse these property expression
but not import expressions or super calls.

## Tests

Added parser snapshot tests for meta and super properties in a new
expression.