[runtime] Fix message of SyntaxError object created from LocalizedParseError (#60)
## Summary

When a SyntaxError object was created from a LocalizedParseError we were
accidentally adding two "SyntaxError:" prefixes to the message. One was
coming from Error.prototype.toString, and the other was coming from the
LocalizedParseError's Display implementation.

We do not want to include this second prefix when creating a SyntaxError
from a LocalizedParseError. Instead we break out a
LocalizedParseError::to_string_without_name method that is used.