[runtime] Error when stack overflows (#64)
## Summary

Let's add an explicit stack overflow check before creating a new stack
frame. This requires calculating the frame size before pushing the
frame. If the stack overflows we throw a RangeError.

Currently we print the entire stack trace in the error, but if this may
be changed in the future it it proves to be too unwieldy.