[runtime] Refactor EvalResult to be an alias for a standard rust Result type (#66)
## Summary

Previously EvalResult was a custom type instead of the standard rust
Result type. This allowed us to write various `impl From<..>`'s to
convert to an EvalResult::Ok from a value using `.into()`. We propagated
errors with the custom `maybe!()` and `must!()` macros, the former of
which is equivalent to the `?` operator.

However the frequent use of `.into()` resulted in a lot of implicit
conversions, and `.into()` to create an `Ok` result is non-idiomatic
rust. We also had to use the more verbose (and non-idomatic) `maybe!()`
macro all over the place. We would like to be more explicit with type
conversions and move closer to normal rust patterns.

This PR switches from a custom EvalResult to an alias of the standard
rust Result. This means that we now need to explicitly wrap all return
values in an `Ok` instead. We also now use the `?` operator instead of
the `maybe!()` macro. However we keep the `must!()` macro for now - the
equivalent would be to unwrap the exception, but keeping `must!()`
clarifies intentions for now.

Instead of `into()` conversion let us also introduce explicit
`as_object()` and `as_value()` conversions for all object values, as
well as `as_value()` conversions for strings and flat strings.

## Tests

No functionality has been changed - tests are unaffected.