[runtime] Use deref operator instead of custom `get()` and `get_()` for handles (#71)
## Summary

Handles have a `get()` or `_get()` method which returns in the inner
value or heap item respectively. However handles already implement the
Deref trait so we can instead use the standard rust deref operator `*`.

Convert all usages of `get() and `_get()` to the deref operator in the
interest of standardizing on common rust patterns.