[regexp] Parse RegExps without named capture groups, then reparase if a named capture group was encountered (#75)
## Summary

Fixes RegExp parsing to happen in two attempts. The first attempt
assumes there are no named capture groups, meaning named backreferences
are parsed as identity escapes and named capture groups cause an error
which signals a named capture group was encountered.

If named capture groups were encountered in the first pass, try parsing
again but this time including named backreferences.

## Tests

Fixes part of test262
`built-ins/String/prototype/split/separator-regexp.js`.

But unfortunately this exposed another error later in the test that
appears to be due to the inclusion of an annex-B feature. This test
includes RegExps with identity escapes for code points in the
ID_Continue Unicode class. Mark it as ignored for now with a note.