[tests] Run test262 tests for all implemented features in CI (#76)
## Summary

With all features implemented except for SharedArrayBuffer, Atomics, and
Unicode Sets, it is past time to start running all test262 tests in CI
to prevent regressions. We now run all test262 tests in CI using
`--ignore-unimplemented` mode. Tests are run in debug, release, and
release with gc stress test modes.

This PR also categorizes all test262 test failures in
`--ignore-unimplemented` mode. We also format the ignored file and
create categories for bugs that we intend to fix, and known divergences
we do not intend to fix.

## Tests

Test262 tests for all implemented features are now running in CI - see
that they pass.