[bytecode] Encode and decode BytecodeSourceMap, adding source positions to errors (#83)
## Summary

Implements mapping of bytecode instructions to source positions. When a
stack trace is printed we use this source mapping to print the line and
column numbers for each frame if possible.

Each BytecodeFunction optionally holds a ByteArray which contains the
encoded source map. The source map consists of a sequence of
(bytecode_offset, source_position) pairs. Each number is encoded as a
varint to minimize space usage using a simple varint encoding: a
sequence of bytes where the high bit signals continuation and the low 7
bits are concatenated to form the number.

BytecodeSourceMaps also hold the source range for the entire function
instead of storing the source range on the BytecodeFunction itself. The
source range is held in the "header" aka the first two varints of the
source map.

The BytecodeSourceMap for a BytecodeFunction is generated alongside the
function. We include the utilities for writing the BytecodeSourceMap but
do not write any entries yet. Note that for efficiency we do not write
the same source position in adjacent entries - we only write a new entry
when there is a new source position.