[runtime] Fix bug when printing error introduced with new top level API (#86)
## Summary

In https://github.com/Hans-Halverson/brimstone/pull/84 we unified all
error types into a BsError and print an error message in main if any
BsError is returned. However we were printing the error after dropping
the Context which is UB. Make sure to only drop the Context after
checking for and printing errors.