[bytecode] Source locations for all binary operations (#88)
## Summary

Write entries to the BytecodeSourceMap for each binary operation
instruction (numeric, bitwise, and comparison). The source position
reported for errors in the binary instruction itself is the position of
the infix binary operator. We must add this position the AST and fill it
during parsing.

Note that we also implement source code mapping for all compound
assignments, which use the same binary operation instructions.

## Tests

Added error snapshot tests for all binary operations and compound
assignments, verifying that reported source position is correct.