[tests] Upgrade to latest test262 (#96)
## Summary

Upgrade to the latest commit of the test262 repo,
8296db887368bbffa3379f995a1e628ddbe8421f. Categorize and suppress all
failing tests. The main remaining area of work here is the large bundle
of `staging/sm` tests which are mass suppressed due to multiple failures
which will require investigation.

Also move stage 4 proposals from the second half of the year into the
unimplemented features section.