[tests] Support raw module tests (#100)
## Summary

Add support to the test runner for raw module tests. We make `is_raw`
its own flag on tests, and allow the test mode to be either a script or
module. Note that the test262 test suite must be reindexed after this
change.

## Tests

This allows us to run the
`/language/module-code/import-attributes/allow-nlt-before-with.js` test.