[regexp] Replace whitespace and digit comparison instructions with generic set comparisons (#105)
## Summary

In preparation for fixes to case insensitive mode we are moving some
special comparison instructions to generic set comparisons. This PR
refactors whitespace and digit comparison instructions to use the
generic set comparison path, meaning individual comparisons are emitted
for each range in the set.

Also fixes generating a set for the unimplemented "Assigned" binary
property.

## Tests

Verify that bytecode snapshot tests for digit and whitespace character
classes are updated with the new comparison instructions.