[regexp] Refactor case insensitive matching to generate case insensitive closures of patterns (#106)
## Summary

Refactors how case insensitive matching is performed in order to fix
case sensitive matching of various sets and prepare for
https://github.com/tc39/proposal-regexp-modifiers. The goals are
avoiding canonicalizing the input string or keeping track of case
sensitivity mode during matching, and instead matching the case
insensitive closure of the pattern.

The Canonicalize abstract operation in non-unicode mode differs slightly
from standard Unicode case folding/mapping procedures. Specifically it
canonicalizes code points using the simple uppercase mapping (instead of
case folding), but avoids mapping any code point from outside the Latin1
range to inside the Latin1 range.

For case insensitive matching we do not canonicalize the input string,
but instead generate the case insensitive closure of the input pattern.
For example each literal in the pattern is replaced with the set of
characters that canonicalize to the same value as that literal (this set
is the case closure).

We use the `add_case_closure_to` method from icu4x to generate the case
closure using case folding. This is always used when in unicode aware
mode. However, in non-unicode mode this method has a handful of code
points that are incorrectly mapped due to the quirks of the Canonicalize
operation. We use a `build.rs` script to generate a module that
precomputes the correct case closures for these code points and refers
to them as case closure overrides. Case insensitive non-unicode matching
then uses these case closure overrides if they exist, and otherwise uses
`add_case_closure_to`.

The one RegExp feature that requires adjustment is backreferences in
case insensitive mode. The input string is no longer canonicalized so we
cannot directly compare two of its slices. Instead we must iterate
through both the captured region and current position and compare
canonicalized values as we go.

## Tests

- Adds RegExp bytecode snapshot tests for simple case insensitive
matching of literals and classes
- Adds RegExp bytecode snapshot tests for case insensitive
backreferences
- Adds RegExp bytecode snapshot tests for case insensitive
- All test262 tests continue to pass