[regexp] Parse non-BMP code points in non-unicode mode RegExps as two code units (#110)
## Summary

Fixes a bug when parsing non-BMP code points in non-unicode mode
RegExps. In non-unicode mode RegExps treat the source string as UTF-16
but treat each surrogate code point individually instead of as a single
code point.

We were doing this correctly on RegExps that are parsed from the heap
but were failing to do this while parsing RegExps from the original
UTF-8 source text. If we are parsing a UTF-8 RegExp we must first
translate it to UTF-16 so that the code units can be parsed
individually.

This also exposes a bug in group name parsing in non-unicode mode. In
non-unicode mode we should parse high surrogate, low surrogate pairs
into a single code point.

## Tests

- Added a parser test that verifies a non-BMP code points is encoded as
individual code units both as a literal and in a class.
- Added a parser test for unicode in group names.
- Fixes 4 test262 unicode sets tests that were failing due to this
behavior