[runtime] Call HasProperty again when loading or storing to object environment (#118)
## Summary

Fixes spec non-compliance exposed by new test262 tests. HasProperty
should be called when loading or storing dynamically from an object
environment in GetBindingValue and SetMutableBinding. In addition we
should error if in strict mode and the property does not still exist at
the HasProperty check.

## Tests

Fixes:
- `language/statements/with/set-mutable-binding-idref-with-proxy-env.js`
- `language/statements/with/get-binding-value-idref-with-proxy-env.js`
- `language/statements/with/set-mutable-binding-idref-with-proxy-env.js`

Keep ignoring the failing
`language/statements/with/set-mutable-binding-idref-compound-assign-with-proxy-env.js`,
which we do not plan on fixing as it fails due to compound assignment
resolving the LHS twice.