[runtime] Lazily generate stack traces on errors (#121)
## Summary

Lazily generate error stack traces instead of generating the entire
stack trace string on error creation. This is accomplished by fetching
the minimal information needed to construct the stack trace on error
creation: an array of stack frames, each of which contains the bytecode
function and current bytecode offset within that function. This array of
stack frames is then attached to the error at creation. When the new
`Error.prototype.stack` getter is called this stack frame info is used
to generate and cache the full stack trace string.

## Tests

All tests continue to pass.