[runtime] Pass new target to rust runtime functions via the stack instead of as native function parameter (#133)
## Summary

The new target object is passed to rust runtime functions as the last
parameter to the native rust function itself. This is not ideal as we
must pass `None` for regular calls and all rust runtime functions
contain an often useless parameter.

Instead let's pass the new target object via the stack for rust runtime
functions. This matches how the new target object is passed for JS
functions and allows us to only pay for the new target when
constructing. Rust runtime functions that are constructors now have a
single register which holds the new target, which is an object if the
new target was passed and otherwise is undefined.

## Tests

All tests continue to pass.