[runtime] Add mode to track handle usage (#135)
## Summary

Add a compile time feature to track usage of handles. Track both the
current number of the handles and max concurrent number of handles that
have been seen. When mode is enabled handle statistics are printed at
the end of the run.

## Tests

Verify that `bs` and test262 runners both work correctly in handle stats
mode.