[runtime] Use HandleScopes to ensure that each VM instruction does not net increase the handle count (#138)
## Summary

We want to avoid gradually increasing the number of handles in use to
avoid allocating too many handle blocks. The largest source remaining is
the VM itself, since most instructions allocate some handles without
deallocating them. To fix this we use handle scopes to ensure that each
instruction does not net increase the handle count.

For instructions that can throw errors this involves wrapping with
`handle_scope!`. Instructions that cannot throw errors use the simpler
`handle_scope_guard!` instead. Instructions that allocate no handles do
not need any changes. Finally we refactor the common call instructions
to only use a handle scope when necessary.

In order to check this we update `handle_stats` mode to assert that the
number of handles before and after execution of the VM are the same.

## Tests

All test262 tests pass, including when in `handle_stats` mode.