[regexp] Fix error source locations in non-unicode mode RegExps (#141)
## Summary

When parsing a source file, non-unicode mode RegExps are first converted
to UTF-16 code units before parsing. However we want to report source
locations in the original UTF-8 file. Previously we just did not report
a source location for these errors, but let's extend the LexerStream
system to track the source file and UTF-16 code unit offset when
reporting errors.

Since errors are reported in a UTF-16 code unit offset we must translate
this to a UTF-8 byte offset, then use this byte offset to determine the
absolute byte position in the UTF-8 input source.

## Tests

Added error snapshot tests for source locations in RegExp parsing
errors, both in unicode and non-unicode mode.