[tests] Refactor test262 runner to generic integration test runner (#145)
## Summary

It is long past time to introduce our own first party integration tests
instead of relying heavily on test262. This PR refactors the test262
runner to be a generic integration test runner that can be used for both
the first party and test262 integration test suites, located in the
`harness` directory. We choose to reuse the metadata and harness files
from test262 for our first party integration tests for ease of use and
for ease of integration if desired.

The new test runner is run with `cargo brimstone-test` which is
implemented as a cargo alias. This will run all test suites by default
but can run a single or set of test suites with the `-s, --suite`
option. If run within the `brimstone` cargo workspace then the runner
will automatically detect the test suites to run by finding the test
manifest. A more detailed summary of the test runner can be found in the
testing README.

The test manifest defines each integration test suite. Each integration
test suite has a root and a path to it's own `ignored_tests.jsonc`.
There is still a central `test_index.json` file which indexes all test
suites.

The test runner now prints a table of test results for all suites. A
generic formatting module was introduced so this can be reused in the
future. The formatter supports text modifiers and text alignment.

## Testing

Added an example first party integration test. Verified this new test
suite runs correctly, and the old test262 test suite runs the exact same
as before.

- Tested with options `--report-test262-progress`, `--save-time-file`,
and `--save-result-files`
- CI passes
- Tested with new `-s, --suite` flag